Check if student is using sklearn.model_selection
Some students are updating the `module sklearn.cross_validation` to the new module `sklearn.model_selection` which has different parameters for the function `ShuffleSplit` (check the [docs](http://scikit-learn.org/stable/modules/generated/sklearn.model_selection.ShuffleSplit.html)). The proposed file change is going to check if the student is using the module `cross_validation`. If `False`, then the `ShuffleSplit` function is updated to the new version of the module.